requirements.txt
chromadb
click>=8.0
pyyaml>=6.0
requests>=2.28
rich>=13.0
